X657C: create a system_ext logical partition when necessary
Usually when building full OTAs for dynamic partition devices the whole partition table is removed and recreated, but in vendorless cases lineageos (and basically any other custom rom) has a patch to disable that functionality and instead just resize the partitions. This, however, breaks the flashing if a new partition is added, because that partition is tried to be resized and, since it doesn't exist, the assert fails. This patch adds another dynamic patch list to create the partition or gracefully fail if it already exists (no assert).

Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>
Signed-off-by: SparXFusion <s2234nadar@gmail.com>